requirements.txt
APScheduler==3.9.1.post1
psycopg2==2.9.5
python-dotenv==0.21.1
python-http-client==3.3.7
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
sendgrid==6.9.7
six==1.16.0
starkbank-ecdsa==2.2.0
tzdata==2022.7
tzlocal==4.2
